# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Question Bank API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List all subjects
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create a subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubjectBody"
      responses:
        "201":
          description: Created subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /subjects/{id}:
    get:
      operationId: getSubject
      tags: [subjects]
      summary: Get a subject by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateSubject
      tags: [subjects]
      summary: Update a subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubjectBody"
      responses:
        "200":
          description: Updated subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSubject
      tags: [subjects]
      summary: Delete a subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # CHAPTERS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Subject:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        chapterCount:
          type: ["integer", "null"]
        questionCount:
          type: ["integer", "null"]
      required:
        - id
        - name
        - createdAt
    CreateSubjectBody:
      type: object
      properties:
        name:
          type: string
      required:
        - name
...